2024_4_15_agent/requirements.txt
openai>=1.0.0
gradio>=4.0.0
python-docx>=0.8.11
PyPDF2>=3.0.0
pandas>=1.5.0